+++ sphinxarg/ext.py
@@ -1,22 +1,43 @@
 from __future__ import annotations
 
 import importlib
+import operator
 import os
 import shutil
 import sys
 from argparse import ArgumentParser
+from typing import TYPE_CHECKING, ClassVar, cast
 
 from docutils import nodes
 from docutils.frontend import get_default_settings
 from docutils.parsers.rst import Parser
 from docutils.parsers.rst.directives import flag, unchanged
 from docutils.statemachine import StringList
+from sphinx.domains import Domain, Index, IndexEntry
+from sphinx.errors import ExtensionError
 from sphinx.ext.autodoc import mock
+from sphinx.roles import XRefRole
+from sphinx.util import logging
 from sphinx.util.docutils import SphinxDirective, new_document
-from sphinx.util.nodes import nested_parse_with_titles
+from sphinx.util.nodes import make_id, make_refnode, nested_parse_with_titles
 
 from sphinxarg import __version__
 from sphinxarg.parser import parse_parser, parser_navigate
+from sphinxarg.utils import command_pos_args, target_to_anchor_id
+
+if TYPE_CHECKING:
+    from collections.abc import Iterable, Sequence
+    from pathlib import Path
+
+    from docutils.nodes import Element
+    from sphinx.addnodes import pending_xref
+    from sphinx.application import Sphinx
+    from sphinx.builders import Builder
+    from sphinx.environment import BuildEnvironment
+
+    _ObjectDescriptionTuple = tuple[str, str, str, str, str, int]
+
+logger = logging.getLogger(__name__)
 
 
 def map_nested_definitions(nested_content):
@@ -283,6 +304,7 @@ def ensure_unique_ids(items):
 
 class ArgParseDirective(SphinxDirective):
     has_content = True
+    required_arguments = 0
     option_spec = {
         'module': unchanged,
         'func': unchanged,
@@ -300,7 +322,9 @@ class ArgParseDirective(SphinxDirective):
         'markdown': flag,
         'markdownhelp': flag,
         'color': flag,
+        'index-groups': unchanged,
     }
+    index_groups: Sequence[str] = ()
 
     def _construct_manpage_specific_structure(self, parser_info):
         """
@@ -423,7 +447,7 @@ class ArgParseDirective(SphinxDirective):
             opt_items = []
             for name in opt['name']:
                 option_declaration = [nodes.option_string(text=name)]
-                if not _is_suppressed(opt['default']):
+                if not self._is_suppressed(opt['default']):
                     option_declaration += nodes.option_argument(
                         '', text='=' + str(opt['default'])
                     )
@@ -486,6 +510,206 @@ class ArgParseDirective(SphinxDirective):
             pass
         # raise exception
         raise FileNotFoundError(self.options['filename'])
+
+    def _print_subcommands(self, data, nested_content, markdown_help=False, settings=None):
+        """
+        Each subcommand is a dictionary with the following keys:
+
+        ['usage', 'action_groups', 'bare_usage', 'name', 'help']
+
+        In essence, this is all tossed in a new section with the title 'name'.
+        Apparently there can also be a 'description' entry.
+        """
+
+        definitions = map_nested_definitions(nested_content)
+        items = []
+        full_subcommand_name_true = self.config.sphinxarg_full_subcommand_name
+        domain = cast('ArgParseDomain', self.env.domains[ArgParseDomain.name])
+
+        if 'children' in data:
+            full_command = command_pos_args(data)
+            node_id = make_id(
+                self.env, self.state.document, '', full_command + '-sub-commands'
+            )
+            target = nodes.target('', '', ids=[node_id])
+            self.set_source_info(target)
+            self.state.document.note_explicit_target(target)
+
+            subcommands = nodes.section(ids=['Sub-commands'])
+            subcommands += nodes.title('Sub-commands', 'Sub-commands')
+
+            for child in data['children']:
+                full_command = command_pos_args(child)
+                node_id = make_id(self.env, self.state.document, '', full_command)
+                target = nodes.target('', '', ids=[node_id])
+                self.set_source_info(target)
+                self.state.document.note_explicit_target(target)
+
+                sec = nodes.section(ids=[node_id, child['name']])
+                if full_subcommand_name_true:
+                    title = nodes.title(full_command, full_command)
+                else:
+                    title = nodes.title(child['name'], child['name'])
+                sec += title
+
+                domain.add_argparse_command(child, node_id, self.index_groups)
+
+                if 'description' in child and child['description']:
+                    desc = [child['description']]
+                elif child['help']:
+                    desc = [child['help']]
+                else:
+                    desc = ['Undocumented']
+
+                # Handle nested content
+                subcontent = []
+                if child['name'] in definitions:
+                    classifier, s, subcontent = definitions[child['name']]
+                    if classifier == '@replace':
+                        desc = [s]
+                    elif classifier == '@after':
+                        desc.append(s)
+                    elif classifier == '@before':
+                        desc.insert(0, s)
+
+                for element in render_list(desc, markdown_help):
+                    sec += element
+                sec += nodes.literal_block(text=child['bare_usage'])
+                for x in self._print_action_groups(
+                    child, nested_content + subcontent, markdown_help, settings=settings
+                ):
+                    sec += x
+
+                for x in self._print_subcommands(
+                    child, nested_content + subcontent, markdown_help, settings=settings
+                ):
+                    sec += x
+
+                if 'epilog' in child and child['epilog']:
+                    for element in render_list([child['epilog']], markdown_help):
+                        sec += element
+
+                subcommands += sec
+            items.append(subcommands)
+
+        return items
+
+    def _print_action_groups(
+        self,
+        data,
+        nested_content,
+        markdown_help=False,
+        settings=None,
+        id_prefix='',
+    ):
+        """
+        Process all 'action groups', which are also include 'Options' and 'Required
+        arguments'. A list of nodes is returned.
+        """
+        definitions = map_nested_definitions(nested_content)
+        nodes_list = []
+        if 'action_groups' in data:
+            for action_group in data['action_groups']:
+                # Every action group is composed of a section, holding
+                # a title, the description, and the option group (members)
+                title_as_id = action_group['title'].replace(' ', '-').lower()
+                full_command = command_pos_args(data)
+                node_id = make_id(
+                    self.env,
+                    self.state.document,
+                    '',
+                    full_command + '-' + title_as_id,
+                )
+                target = nodes.target('', '', ids=[node_id])
+                self.set_source_info(target)
+                self.state.document.note_explicit_target(target)
+
+                section = nodes.section(ids=[node_id, f'{id_prefix}-{title_as_id}'])
+                section += nodes.title(action_group['title'], action_group['title'])
+
+                desc = []
+                if action_group['description']:
+                    desc.append(action_group['description'])
+                # Replace/append/prepend content to the description according to nested content
+                subcontent = []
+                if action_group['title'] in definitions:
+                    classifier, s, subcontent = definitions[action_group['title']]
+                    if classifier == '@replace':
+                        desc = [s]
+                    elif classifier == '@after':
+                        desc.append(s)
+                    elif classifier == '@before':
+                        desc.insert(0, s)
+                    elif classifier == '@skip':
+                        continue
+                    if len(subcontent) > 0:
+                        for k, v in map_nested_definitions(subcontent).items():
+                            definitions[k] = v
+                # Render appropriately
+                for element in render_list(desc, markdown_help):
+                    section += element
+
+                local_definitions = definitions
+                if len(subcontent) > 0:
+                    local_definitions = dict(definitions.items())
+                    for k, v in map_nested_definitions(subcontent).items():
+                        local_definitions[k] = v
+
+                items = []
+                # Iterate over action group members
+                for entry in action_group['options']:
+                    # Members will include:
+                    #    default	The default value. This may be ==SUPPRESS==
+                    #    name	A list of option names (e.g., ['-h', '--help']
+                    #    help	The help message string
+                    # There may also be a 'choices' member.
+                    # Build the help text
+                    arg = []
+                    if 'choices' in entry:
+                        arg.append(
+                            f'Possible choices: {", ".join(map(str, entry["choices"]))}\n'
+                        )
+                    if 'help' in entry:
+                        arg.append(entry['help'])
+                    if not self._is_suppressed(entry['default']):
+                        # Put the default value in a literal block,
+                        # but escape backticks already in the string
+                        default_str = str(entry['default']).replace('`', r'\`')
+                        arg.append(f'Default: ``{default_str}``')
+
+                    # Handle nested content, the term used in the dict
+                    # has the comma removed for simplicity
+                    desc = arg
+                    term = ' '.join(entry['name'])
+                    if term in local_definitions:
+                        classifier, s, subcontent = local_definitions[term]
+                        if classifier == '@replace':
+                            desc = [s]
+                        elif classifier == '@after':
+                            desc.append(s)
+                        elif classifier == '@before':
+                            desc.insert(0, s)
+                    term = ', '.join(entry['name'])
+
+                    n = nodes.option_list_item(
+                        '',
+                        nodes.option_group('', nodes.option_string(text=term)),
+                        nodes.description('', *render_list(desc, markdown_help, settings)),
+                    )
+                    items.append(n)
+
+                section += nodes.option_list('', *items)
+                nodes_list.append(section)
+
+        return nodes_list
+
+    @staticmethod
+    def _is_suppressed(item: str | None) -> bool:
+        """Return whether item should not be printed."""
+        if item is None:
+            return True
+        item = str(item).replace('"', '').replace("'", '')
+        return item == '==SUPPRESS=='
 
     def run(self):
         if 'module' in self.options and 'func' in self.options:
@@ -573,9 +797,25 @@ class ArgParseDirective(SphinxDirective):
                 items.extend(render_list([result['description']], True))
             else:
                 items.append(self._nested_parse_paragraph(result['description']))
+
+        if 'index-groups' in self.options:
+            self.index_groups = list(map(str.strip, self.options['index-groups'].split(', ')))
+        else:
+            self.index_groups = []
+
+        full_command = command_pos_args(result)
+        node_id = make_id(self.env, self.state.document, '', full_command)
+        target = nodes.target('', '', ids=[node_id])
+        items.append(target)
+        self.set_source_info(target)
+        self.state.document.note_explicit_target(target)
+
+        domain = cast('ArgParseDomain', self.env.get_domain(ArgParseDomain.name))
+        domain.add_argparse_command(result, node_id, self.index_groups)
+
         items.append(nodes.literal_block(text=result['usage']))
         items.extend(
-            print_action_groups(
+            self._print_action_groups(
                 result,
                 nested_content,
                 markdown_help,
@@ -585,7 +825,7 @@ class ArgParseDirective(SphinxDirective):
         )
         if 'nosubcommands' not in self.options:
             items.extend(
-                print_subcommands(
+                self._print_subcommands(
                     result,
                     nested_content,
                     markdown_help,
@@ -601,9 +841,188 @@ class ArgParseDirective(SphinxDirective):
         return items
 
 
-def setup(app):
+class CommandsIndex(Index):
+    name = 'index'
+    localname = 'Commands Index'
+
+    def generate(
+        self, docnames: Iterable[str] | None = None
+    ) -> tuple[list[tuple[str, list[IndexEntry]]], bool]:
+        content: dict[str, list[IndexEntry]] = {}
+        commands: list[_ObjectDescriptionTuple]
+        commands = sorted(self.domain.get_objects(), key=operator.itemgetter(0))
+        for cmd, dispname, _typ, docname, anchor, priority in commands:
+            inx_entry = IndexEntry(cmd, priority, docname, anchor, docname, '', dispname)
+            content.setdefault(cmd[0].lower(), []).append(inx_entry)
+        return sorted(content.items()), True
+
+
+class CommandsByGroupIndex(Index):
+    # Defaults (can be overridden through `conf.py`):
+    name = 'by-group'
+    localname = 'Commands by Group'
+
+    def generate(
+        self, docnames: Iterable[str] | None = None
+    ) -> tuple[list[tuple[str, list[IndexEntry]]], bool]:
+        content: dict[str, list[IndexEntry]] = {}
+        commands_by_group: dict[str, list[_ObjectDescriptionTuple]]
+        commands_by_group = self.domain.data['commands-by-group']
+        for group in sorted(commands_by_group):
+            commands = sorted(commands_by_group[group], key=operator.itemgetter(0))
+            for cmd, dispname, _typ, docname, anchor, priority in commands:
+                idx_entry = IndexEntry(cmd, priority, docname, anchor, docname, '', dispname)
+                content.setdefault(group, []).append(idx_entry)
+        return sorted(content.items()), True
+
+
+class ArgParseDomain(Domain):
+    name = 'commands'
+    label = 'commands-label'
+
+    roles = {
+        'command': XRefRole(),
+    }
+    indices = []
+    initial_data: ClassVar[
+        dict[str, list[_ObjectDescriptionTuple] | dict[str, list[_ObjectDescriptionTuple]]]
+    ] = {
+        'commands': [],
+        'commands-by-group': {},
+    }
+
+    # Keep a list of the temporary index files that are created in the
+    # source directory. The files are created if the command_xxx_in_toctree
+    # option is set to True.
+    temporary_index_files: list[Path] = []
+
+    def get_objects(self) -> Iterable[_ObjectDescriptionTuple]:
+        yield from self.data['commands']
+
+    def resolve_xref(
+        self,
+        env: BuildEnvironment,
+        fromdocname: str,
+        builder: Builder,
+        typ: str,
+        target: str,
+        node: pending_xref,
+        contnode: Element,
+    ) -> nodes.reference | None:
+        anchor_id = target_to_anchor_id(target)
+        match = [
+            (docname, anchor)
+            for _cmd, _sig, _type, docname, anchor, _prio in self.get_objects()
+            if anchor_id == anchor
+        ]
+
+        if len(match) > 0:
+            todocname = match[0][0]
+            targ = match[0][1]
+
+            return make_refnode(builder, fromdocname, todocname, targ, contnode, targ)
+        else:
+            msg = f'Error, no command xref target from {fromdocname}:{target}'
+            logger.warning(msg)
+            return None
+
+    def add_argparse_command(self, result: dict, anchor: str, groups: Sequence[str] = ()):
+        """Add an argparse command to the domain."""
+        full_command = command_pos_args(result)
+        desc = result.get('description', 'No description.')
+        idx_entry = (full_command, desc, 'command', self.env.docname, anchor, 0)
+        self.data['commands'].append(idx_entry)
+
+        # A likely duplicate list of index entries is kept for the grouping.
+        # A separate list is kept to avoid the edge case that a command is used
+        # once as part of a group (with index_groups) and another time without the
+        # option.
+        commands_by_group = self.data['commands-by-group']
+        for group in groups:
+            commands_by_group.setdefault(group, []).append(idx_entry)
+
+
+def _delete_temporary_files(app: Sphinx, _err) -> None:
+    assert app.env is not None
+    domain = cast('ArgParseDomain', app.env.domains[ArgParseDomain.name])
+    for fpath in domain.temporary_index_files:
+        fpath.unlink(missing_ok=True)
+
+
+def _create_temporary_dummy_file(
+    app: Sphinx, domain: Domain, docname: str, title: str
+) -> None:
+    dummy_file = app.srcdir / docname
+    if dummy_file.exists():
+        msg = (
+            f'The Sphinx project cannot include a file named '
+            f'"{docname}" in the source directory.'
+        )
+        raise ExtensionError(msg)
+
+    underline = len(title) * '='
+    content = '\n'.join((
+        f'{title}',
+        f'{underline}',
+        '',
+        'Temporary file that is replaced with an index from the sphinxarg extension.',
+        f'Creating this temporary file enables you to add {docname} to the toctree.',
+    ))
+    dummy_file.write_text(content, encoding='utf-8')
+    domain = cast('ArgParseDomain', domain)
+    domain.temporary_index_files.append(dummy_file)
+
+
+def configure_ext(app: Sphinx) -> None:
+    domain = cast('ArgParseDomain', app.env.domains[ArgParseDomain.name])
+    build_index = False
+    build_by_group_index = False
+
+    CommandsByGroupIndex.name = app.config.sphinxarg_commands_by_group_index_file_suffix
+    CommandsByGroupIndex.localname = app.config.sphinxarg_commands_by_group_index_title
+
+    if app.config.sphinxarg_commands_index_in_toctree:
+        build_index = True
+        docname = f'{ArgParseDomain.name}-{CommandsIndex.name}.rst'
+        _create_temporary_dummy_file(app, domain, docname, CommandsIndex.localname)
+
+    if app.config.sphinxarg_commands_by_group_index_in_toctree:
+        build_by_group_index = True
+        docname = f'{ArgParseDomain.name}-{CommandsByGroupIndex.name}.rst'
+        _create_temporary_dummy_file(app, domain, docname, CommandsByGroupIndex.localname)
+
+    if build_index or app.config.sphinxarg_build_commands_index:
+        domain.indices.append(CommandsIndex)
+
+    if build_by_group_index or app.config.sphinxarg_build_commands_by_group_index:
+        domain.indices.append(CommandsByGroupIndex)
+
+    # Call setup so that :ref:`commands-...` are link targets.
+    domain.setup()
+
+
+def setup(app: Sphinx):
     app.setup_extension('sphinx.ext.autodoc')
+    app.add_domain(ArgParseDomain)
     app.add_directive('argparse', ArgParseDirective)
+
+    # Config options must be mentioned in ``usage.rst`` too!
+
+    app.add_config_value('sphinxarg_full_subcommand_name', False, 'html', bool)
+    app.add_config_value('sphinxarg_build_commands_index', False, 'html', bool)
+    app.add_config_value('sphinxarg_commands_index_in_toctree', False, 'html', bool)
+
+    app.add_config_value('sphinxarg_build_commands_by_group_index', False, 'html', bool)
+    app.add_config_value('sphinxarg_commands_by_group_index_in_toctree', False, 'html', bool)
+    app.add_config_value(
+        'sphinxarg_commands_by_group_index_file_suffix', CommandsByGroupIndex.name, 'html', str
+    )
+    app.add_config_value(
+        'sphinxarg_commands_by_group_index_title', CommandsByGroupIndex.localname, 'html', str
+    )
+
+    app.connect('builder-inited', configure_ext)
+    app.connect('build-finished', _delete_temporary_files)
     return {
         'version': __version__,
         'parallel_read_safe': True,