+++ test/roots/test-command-by-group-index/conf.py
@@ -1,0 +1,2 @@
+extensions = ['sphinxarg.ext']
+sphinxarg_commands_by_group_index_in_toctree = True