+++ test/roots/test-command-index/conf.py
@@ -1,0 +1,3 @@
+extensions = ['sphinxarg.ext']
+sphinxarg_build_commands_index = True
+sphinxarg_commands_index_in_toctree = True