+++ test/test_argparse_directive.py
@@ -1,0 +1,8 @@
+import pytest
+
+
+@pytest.mark.skip(reason='Refactoring')
+@pytest.mark.sphinx('html', testroot='argparse-directive')
+def test_bad_index_groups(app, status, warning):
+    app.build()
+    assert 'failed to parse index-groups as a list' in warning.getvalue()