+++ test/test_parser.py
@@ -138,6 +138,10 @@ def test_parse_nested():
                     ],
                 },
             ],
+            'parent': {
+                'name': '',
+                'prog': 'under-test',
+            },
         }
     ]
 
@@ -185,6 +189,10 @@ def test_parse_nested_with_alias():
                     ],
                 },
             ],
+            'parent': {
+                'name': '',
+                'prog': 'under-test',
+            },
         }
     ]
 
@@ -205,6 +213,10 @@ def test_aliased_traversal():
         'usage': 'usage: under-test level1 [-h]',
         'name': 'level1 (l1)',
         'identifier': 'level1',
+        'parent': {
+            'name': '',
+            'prog': 'under-test',
+        },
     }
 
 
@@ -232,6 +244,19 @@ def test_parse_nested_traversal():
         {'name': ['bar'], 'help': '', 'default': None},
     ]
 
+    assert data3['parent'] == {
+        'name': 'level2',
+        'prog': '',
+        'parent': {
+            'name': 'level1',
+            'prog': '',
+            'parent': {
+                'name': '',
+                'prog': 'under-test',
+            },
+        },
+    }
+
     data2 = parser_navigate(data, 'level1 level2')
     assert data2['children'] == [
         {
@@ -249,10 +274,20 @@ def test_parse_nested_traversal():
                     ],
                 }
             ],
+            'parent': {
+                'name': 'level2',
+                'prog': '',
+                'parent': {
+                    'name': 'level1',
+                    'prog': '',
+                    'parent': {'name': '', 'prog': 'under-test'},
+                },
+            },
         }
     ]
 
     assert data == parser_navigate(data, '')
+    assert 'parent' not in data
 
 
 def test_fill_in_default_prog():
@@ -387,6 +422,10 @@ def test_action_groups_with_subcommands():
             'bare_usage': 'foo A [-h] baz',
             'name': 'A',
             'help': 'A subparser',
+            'parent': {
+                'name': '',
+                'prog': 'foo',
+            },
         },
         {
             'usage': 'usage: foo B [-h] [--barg {X,Y,Z}]',
@@ -407,5 +446,9 @@ def test_action_groups_with_subcommands():
             'bare_usage': 'foo B [-h] [--barg {X,Y,Z}]',
             'name': 'B',
             'help': 'B subparser',
+            'parent': {
+                'name': '',
+                'prog': 'foo',
+            },
         },
     ]